refactor: 세팅룸 미디어스트림 관리 hook 분리
